[FIX] 브랜치 merge 작업에 따른 디버깅 및 수정
