+++ src/lib/api.js
@@ -1,6 +1,6 @@
 import { getAuthToken, isTokenExpired } from './auth'
 
-const API_BASE_URL = 'http://localhost:8080/api'
+const API_BASE_URL = `${import.meta.env.VITE_SERVER_URL}/api`
 
 
 // 기본 fetch 래퍼
